polyhal_trap::trap

Function uservec

#[no_mangle]
pub unsafe extern "C" fn uservec()

User Trap Entry

This function is called when a user process is interrupted. It saves the user context and switches to kernel mode.

Safety

This function is unsafe because it performs low-level operations that can lead to undefined behavior if not used correctly.